Reduce indexer tests to new scope
